added documentation to start project frontend deleted console logs for booking

+++ graphql/resolvers/booking.js
@@ -37,15 +37,13 @@ module.exports =  {
         }
     },
     cancelBooking: async (args, req) => {
-        // if (!req.isAuth) {
-        //   throw new Error('Unauthenticated!');
-        // }
+        if (!req.isAuth) {
+          throw new Error('Unauthenticated!');
+        }
         try {
-          const booking = await Booking.findById(args.bookingId).populate('events');
-          console.log("BOOKING", booking);
+          const booking = await Booking.findById(args.bookingId).populate('event');
           const event = transformEvent(booking.event);
           await Booking.deleteOne({ _id: args.bookingId });
-          console.log("EVENT", event);
           return event;
         } catch (err) {
           throw err;
